feat: replace GSIGEO2024 beta file with JPGEO2024
