+++ CMakeLists.txt
@@ -29,7 +29,7 @@ link_directories(
 add_library(llh_converter SHARED
   src/height_converter.cpp
   src/gsigeo.cpp
-  src/gsigeo2024.cpp
+  src/jpgeo2024.cpp
   src/llh_converter.cpp
   src/meridian_convergence_angle_correction.cpp
 )
@@ -48,9 +48,9 @@ add_executable(gsigeo_test test/gsigeo_test.cpp)
 target_link_libraries(gsigeo_test ${catkin_LIBRARIES} llh_converter)
 add_dependencies(gsigeo_test ${catkin_EXPORTED_TARGETS})
 
-add_executable(gsigeo2024_test test/gsigeo2024_test.cpp)
-target_link_libraries(gsigeo2024_test ${catkin_LIBRARIES} llh_converter)
-add_dependencies(gsigeo2024_test ${catkin_EXPORTED_TARGETS})
+add_executable(jpgeo2024_test test/jpgeo2024_test.cpp)
+target_link_libraries(jpgeo2024_test ${catkin_LIBRARIES} llh_converter)
+add_dependencies(jpgeo2024_test ${catkin_EXPORTED_TARGETS})
 
 install(DIRECTORY include/llh_converter/
   DESTINATION include/llh_converter
