+++ README.md
@@ -17,7 +17,7 @@ Convert height between ellipsoid and orthometric library
 
 * EGM2008-1
 * GSIGEO2011 Ver2.1
-* GSIGEO2024beta (Data file will be updated in 2025/04)
+* JPGEO2024
 
 ### Usage
 
@@ -125,19 +125,19 @@ cd test_ws/src/
 git clone https://github.com/MapIV/llh_converter.git
 sudo mkdir /usr/share/GSIGEO
 sudo cp llh_converter/data/gsigeo2011_ver2_1.asc /usr/share/GSIGEO/
-unzip llh_converter/data/GSIGEO2024beta.zip
-sudo mv GSIGEO2024beta.isg /usr/share/GSIGEO/
+unzip llh_converter/data/JPGEO2024.zip
+sudo mv JPGEO2024.isg /usr/share/GSIGEO/
 cd ../
 catkin_make -DCMAKE_BUILD_TYPE=Release
 ```
 
 ## Geoid model data
 
-This package contains GSIGEO2011/GSIGEO2024 geoid data file which is provided by Geospatial Information Authority of Japan.
+This package contains GSIGEO2011/JPGEO2024 geoid data file which is provided by Geospatial Information Authority of Japan.
 
 [GSI's official website that relies on GSIGEO2011](https://fgd.gsi.go.jp/download/geoid.php)
 
-[GSI's official website that relies on GSIGEO2024](https://www.gsi.go.jp/buturisokuchi/grageo_reference.html)
+[GSI's official website that relies on JPGEO2024](https://www.gsi.go.jp/buturisokuchi/grageo_reference.html)
 
 ## LICENSE
