--- include/llh_converter/gsigeo2024.hpp
@@ -1,58 +1,0 @@
-// BSD 3-Clause License
-//
-// Copyright (c) 2025, Map IV, Inc.
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-// 1. Redistributions of source code must retain the above copyright notice, this
-//    list of conditions and the following disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright notice,
-//    this list of conditions and the following disclaimer in the documentation
-//    and/or other materials provided with the distribution.
-//
-// 3. Neither the name of the copyright holder nor the names of its
-//    contributors may be used to endorse or promote products derived from
-//    this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-#ifndef M4_GSIGEO2024_HPP
-#define M4_GSIGEO2024_HPP
-
-#include <string>
-#include <vector>
-
-namespace llh_converter
-{
-class GSIGEO2024
-{
-public:
-  GSIGEO2024();
-  ~GSIGEO2024();
-  void loadGeoidMap(const std::string& geoid_file);
-  double getGeoid(const double& lat, const double& lon);
-
-private:
-  // Geoid data
-  std::vector<std::vector<double>> geoid_map_;  // geoid_map_[lat][lon]
-  int column_size_ = 1601;
-  int row_size_ = 2101;
-
-  // Flag
-  bool is_geoid_loaded_ = false;
-};
-}  // namespace llh_converter
-
-#endif
