+++ include/llh_converter/height_converter.hpp
@@ -36,7 +36,7 @@
 #include <GeographicLib/Geoid.hpp>
 
 #include "llh_converter/gsigeo.hpp"
-#include "llh_converter/gsigeo2024.hpp"
+#include "llh_converter/jpgeo2024.hpp"
 
 namespace llh_converter
 {
@@ -44,7 +44,7 @@ enum class GeoidType
 {
   EGM2008 = 0,
   GSIGEO2011 = 1,
-  GSIGEO2024 = 2,
+  JPGEO2024 = 2,
   NONE = 3,
 };
 
@@ -68,8 +68,8 @@ public:
   double getGeoidDeg(const double& lat_deg, const double& lon_deg);
   void loadGSIGEOGeoidFile(const std::string& geoid_file);
   void loadGSIGEOGeoidFile();
-  void loadGSIGEO2024GeoidFile(const std::string& geoid_file);
-  void loadGSIGEO2024GeoidFile();
+  void loadJPGEO2024GeoidFile(const std::string& geoid_file);
+  void loadJPGEO2024GeoidFile();
 
 private:
   // Geoid Type
@@ -80,11 +80,11 @@ private:
   // Geoid maps
   GeographicLib::Geoid egm2008_{ "egm2008-1" };
   GSIGEO2011 gsigeo2011_;
-  GSIGEO2024 gsigeo2024_;
+  JPGEO2024 jpgeo2024_;
 
   double getGeoidEGM2008(const double& lat_rad, const double& lon_rad);
   double getGeoidGSIGEO2011(const double& lat_rad, const double& lon_rad);
-  double getGeoidGSIGEO2024(const double& lat_rad, const double& lon_rad);
+  double getGeoidJPGEO2024(const double& lat_rad, const double& lon_rad);
 };
 }  // namespace llh_converter
